// src/types/templateCreation/templateErrors.ts
export enum CREATE_TEMPLATE_ERROR_TEXTS {
    NAME_EMPTY = "name is empty",
    HTML_CONTENT_EMPTY = "htmlContent is empty",
    TEXT_CONTENT_EMPTY = "textContent is empty",
    EXAMPLE_HTML_CONTENT_EMPTY = "exampleHtmlContent is empty",
    EXAMPLE_TEXT_CONTENT_EMPTY = "exampleTextContent is empty",
    HTML_CONTENT_INVALID = "htmlContent is invalid",
    TEXT_CONTENT_INVALID = "textContent is invalid",
    EXAMPLE_HTML_CONTENT_INVALID = "exampleHtmlContent is invalid",
    EXAMPLE_TEXT_CONTENT_INVALID = "exampleTextContent is invalid",
}